x/swap/types/message_distribute_coins_test.go
package types

import (
	"testing"

	"coswap/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgDistributeCoins_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDistributeCoins
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDistributeCoins{
				Admin: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDistributeCoins{
				Admin: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
